completed access control implementation

+++ src/main/java/seedu/address/logic/commands/DeleteCommand.java
@@ -24,6 +24,8 @@ public class DeleteCommand extends Command {
 
     public static final String MESSAGE_DELETE_PERSON_SUCCESS = "Deleted Person: %1$s";
 
+    public static final String MESSAGE_DELETE_ERROR = "You cannot delete your own account!";
+
     private final Index targetIndex;
 
     public DeleteCommand(Index targetIndex) {
@@ -33,6 +35,8 @@ public class DeleteCommand extends Command {
     @Override
     public CommandResult execute(Model model) throws CommandException {
         requireNonNull(model);
+
+        // Access Control check for command operation
         if (!Person.isAdmin(model.getLoggedInPerson())) {
             throw new CommandException(Messages.MESSAGE_ACCESS_ADMIN);
         }
@@ -44,6 +48,12 @@ public class DeleteCommand extends Command {
         }
 
         Person personToDelete = lastShownList.get(targetIndex.getZeroBased());
+
+        //Prevent deletion of own account
+        if (personToDelete.equals(model.getLoggedInPerson())) {
+            throw new CommandException(MESSAGE_DELETE_ERROR);
+        }
+
         model.deletePerson(personToDelete);
         return new CommandResult(String.format(MESSAGE_DELETE_PERSON_SUCCESS, personToDelete));
     }
